Add files via upload
the first look of the ongoing tinct vision! in this code its going to recognize the colour by the most dominant color that shows in the cam. also when hand is shown in the camera, the method of recognize the colour by dominant in the cam is turn off then its only detects colour on what the hand is holding... by now, the colour detection when switch to the hand its likely still broken because its hard to trace outside the hand, aka invert the colour outside the hand.